requirements.txt
cryptography==44.0.2
platformdirs==4.3.8
termcolor==3.1.0
requests==2.32.3
pywebview==5.4
flask==3.1.1